+++ src/ytsimpledownloader/__init__.py
@@ -1,3 +1,3 @@
-"""Single-video YouTube downloader."""
+"""Simple YouTube downloader."""
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"